app/src/main/java/com/dicoding/moviecatalogue/ui/detail/tvshow/DetailTvShowViewModel.kt
package com.dicoding.moviecatalogue.ui.detail.tvshow

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.dicoding.moviecatalogue.data.DataRepository
import com.dicoding.moviecatalogue.data.source.local.TvShowEntity

class DetailTvShowViewModel(private val dataRepository: DataRepository) : ViewModel() {
    private lateinit var tvShowId: String

    fun setSelectedTvShow(tvShowId: String) {
        this.tvShowId = tvShowId
    }

    fun getTvShow(): LiveData<TvShowEntity> = dataRepository.getTvShowById(tvShowId)

}